added: Google auth key to local.properties and its used with buildconfig so buildconfig settings added to AuthenticationPlugin
